+++ db/migrate/003_add_timestamps_to_rates.rb
@@ -1,0 +1,5 @@
+class AddTimestampsToRates < ActiveRecord::Migration
+  def change
+    add_timestamps(:rates, :null => false)
+  end
+end