+++ models/candle_stick.rb
@@ -1,0 +1,2 @@
+class CandleStick < ActiveRecord::Base
+end