+++ CHANGELOG.md
@@ -1,2 +1,28 @@
+# 2.1.5 (2018/12/23)
+- [UPDATE] gems
+
+# 2.1.4 (2018/12/22)
+- [UPDATE] gems
+
+# 2.1.3 (2018/11/14)
+- [FIX] file check when removing old files
+- [ADD] created_at and updated_at to rates table
+
+# 2.1.2 (2018/11/04)
+- [FIX] bug of script to compress backup data
+
+# 2.1.1 (2018/10/30)
+- [FIX] bug of script to import rates and candlr sticks
+
+# 2.1.0 (2018/10/29)
+- [FIX] bug of script to remove files
+- [ADD] arguments to specify period to import
+
+# 2.0.1 (2018/10/27)
+- [UPDATE] tool to remove candle sticks
+
+# 2.0.0 (2018/10/27)
+- [ADD] script to import candle sticks
+
 # 1.0.0 (2018/08/22)
 - [NEW] create app